+++ toolium/test/utils/test_path_utils.py
@@ -26,7 +26,7 @@ import uuid
 
 import pytest
 
-from toolium.path_utils import get_valid_filename, makedirs_safe
+from toolium.utils.path_utils import get_valid_filename, makedirs_safe
 
 filename_tests = (
     ('hola_pepito', 'hola_pepito'),
@@ -90,3 +90,20 @@ def test_create_new_folder_parallel():
         assert exceptions.qsize() == 0
         assert os.path.isdir(folder)
         os.rmdir(folder)
+
+
+def test_path_utils_filename_compatibility():
+    # Check that path_utils works with old imports
+    from toolium.path_utils import get_valid_filename
+
+    assert 'test' == get_valid_filename('test')
+
+
+def test_path_utils_makedirs_compatibility():
+    # Check that path_utils works with old imports
+    from toolium.path_utils import makedirs_safe
+
+    folder = os.path.join('output', str(uuid.uuid4()))
+    makedirs_safe(folder)
+    assert os.path.isdir(folder)
+    os.rmdir(folder)