+++ CHANGELOG.rst
@@ -6,6 +6,8 @@ v1.7.2
 
 *In development*
 
+- Move utils.py and path_utils.py files to utils folder maintaining backwards compatibility
+
 v1.7.1
 ------